backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
Pillow~=9.2.0
django-cors-headers==3.13.0
djoser==2.1.0
psycopg2==2.9.5
drf-spectacular==0.26.4
drf-spectacular[sidecar]
python-dotenv~=0.21.0
gunicorn==20.1.0
python-json-logger==2.0.6
pytest-django==4.5.2
openpyxl==3.1.2
django-filter==22.1